Move classes to individual files. General reorganization of code.
